backend/requirements.txt
Django>=5.2
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt